feat(ui): add mobile ttabs root
